Make Keypad a functional component (#2054)
