+++ app/components/Base/Keypad/rules/index.js
@@ -1,0 +1,9 @@
+import native from './native';
+import usd from './usd';
+
+const rules = {
+	native,
+	usd
+};
+
+export default rules;
